requirements.txt
feedparser==6.0.11
google-generativeai==0.7.2
requests==2.32.3
yt-dlp==2024.11.18
